determine queryBuilderType if called on other method
